Trends service

+++ trends-service/src/main/java/ar/edu/utn/ba/ddsi/smartlife/trends_service/dtos/evento/VentaRegistradaEvento.java
@@ -2,6 +2,6 @@ package ar.edu.utn.ba.ddsi.smartlife.trends_service.dtos.evento;
 
 public record VentaRegistradaEvento(
 	long productoId,
-	long cantidadVendida
+	int cantidadVendida
 ) {
 }
